requirements.txt
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==4.1.2
billiard==4.2.4
boto3==1.34.34
botocore==1.34.162
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorlog==6.10.1
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.109.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
jmespath==1.0.1
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
minio==7.2.0
packaging==25.0
passlib==1.7.4
pillow==12.1.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.6.0
pydantic_core==2.16.1
PyMuPDF==1.26.7
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.6
redis==5.0.1
rsa==4.9.1
s3transfer==0.10.4
six==1.17.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.27.0
vine==5.1.0
wcwidth==0.2.14
structlog
pytest
httpx